pyproject.toml
[project]
name = "djangoproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.12.0",
    "celery>=5.5.3",
    "celery-types==0.23.0",
    "django>=5.2.4",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "djoser>=2.3.3",
    "drf-spectacular>=0.28.0",
    "drf-yasg>=1.21.10",
    "flower>=2.0.1",
    "ipython>=9.4.0",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "ruff>=0.12.3",
]

[dependency-groups]
prod = [
    "psycopg[binary]>=3.2.9",
    "uwsgi>=2.0.30",
]
